requirements.txt
streamlit
streamlit-chat
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
numpy
python-dotenv



